SERVER-7273 Use static_cast when passing long long to Value::createDouble().

+++ src/mongo/dbtests/documenttests.cpp
@@ -805,7 +805,7 @@ namespace DocumentTests {
             class LongToDouble : public ToDoubleBase {
                 intrusive_ptr<const Value> value() {
                     // A long that cannot be exactly represented as a double.
-                    return Value::createDouble( 0x8fffffffffffffffLL );
+                    return Value::createDouble( static_cast<double>( 0x8fffffffffffffffLL ) );
                 }
                 double expected() { return static_cast<double>( 0x8fffffffffffffffLL ); }
             };
